dockerLab/fig_codelab/requirements.txt
Django
south
djangorestframework
markdown
django-filter
django-storages
gunicorn
virtualenv
